docker/producer/requirements.txt
numpy
apache_beam
beam_nuggets
kafka-python

